# Resolved Conflicts:
#	src/TrueOrFalse.Frontend.Web/Views/Users/Network.aspx

+++ src/TrueOrFalse/Domain/Category/History/EditData/CategoryRelation_EditData_V2.cs
@@ -5,7 +5,7 @@ public class CategoryRelation_EditData_V2 : CategoryRelation_EditData
     public CategoryRelation_EditData_V2(CategoryRelation categoryRelation)
     {
         CategoryId = categoryRelation.Category.Id;
-        RelationType = (int)categoryRelation.CategoryRelationType;
+        RelationType = categoryRelation.CategoryRelationType;
         RelatedCategoryId = (int)categoryRelation.RelatedCategory.Id;
     }
 }
